requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
python-jose
flask-cors
six
python-dotenv
requests
authlib
requests_oauthlib
